➕ Get user route and method

+++ migrations/2025-04-09-131347_users/up.sql
@@ -2,6 +2,6 @@
 
 CREATE TABLE users (
     id INTEGER AUTO_INCREMENT PRIMARY KEY,
-    discord_id INTEGER NOT NULL,
+    discord_id BIGINT NOT NULL,
     anilist_username VARCHAR(50) NOT NULL
 );
